backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# CORS
fastapi-cors==0.0.6

# Validation
email-validator==2.1.0